Reorganize dynamic schema/dto logic into separate directory.
