+++ resources/ApiKey.py
@@ -3,7 +3,7 @@ from flask_restx import fields
 
 from middleware.primary_resource_logic.login_queries import get_api_key_for_user
 from middleware.primary_resource_logic.user_queries import UserRequest
-from middleware.schema_and_dto_logic.model_helpers_with_schemas import create_user_model
+from middleware.schema_and_dto_logic.dynamic_logic.model_helpers_with_schemas import create_user_model
 from utilities.namespace import create_namespace, AppNamespaces
 
 from resources.PsycopgResource import PsycopgResource, handle_exceptions
