+++ resources/DataSources.py
@@ -17,16 +17,15 @@ from middleware.decorators import (
 from middleware.primary_resource_logic.data_sources_logic import (
     get_data_sources_wrapper,
     data_source_by_id_wrapper,
-    get_data_sources_for_map_wrapper,
     add_new_data_source_wrapper,
     update_data_source_wrapper,
     DataSourcesGetManyRequestDTO,
     delete_data_source_wrapper,
 )
-from middleware.schema_and_dto_logic.dynamic_schema_documentation_construction import (
+from middleware.schema_and_dto_logic.dynamic_logic.dynamic_schema_documentation_construction import (
     get_restx_param_documentation,
 )
-from middleware.schema_and_dto_logic.model_helpers_with_schemas import (
+from middleware.schema_and_dto_logic.dynamic_logic.model_helpers_with_schemas import (
     CRUDModels,
 )
 from middleware.schema_and_dto_logic.primary_resource_schemas.data_sources_schemas import DataSourcesGetByIDSchema, \
