+++ resources/Login.py
@@ -4,7 +4,7 @@ from config import limiter
 from middleware.primary_resource_logic.login_queries import try_logging_in
 from middleware.primary_resource_logic.user_queries import UserRequest
 from resources.resource_helpers import create_jwt_tokens_model
-from middleware.schema_and_dto_logic.model_helpers_with_schemas import create_user_model
+from middleware.schema_and_dto_logic.dynamic_logic.model_helpers_with_schemas import create_user_model
 from utilities.namespace import create_namespace
 
 from resources.PsycopgResource import PsycopgResource, handle_exceptions
