+++ resources/Permissions.py
@@ -13,7 +13,7 @@ from middleware.primary_resource_logic.permissions_logic import (
 from resources.PsycopgResource import handle_exceptions, PsycopgResource
 from resources.resource_helpers import add_jwt_header_arg
 from utilities.namespace import AppNamespaces, create_namespace
-from middleware.schema_and_dto_logic.dynamic_schema_documentation_construction import (
+from middleware.schema_and_dto_logic.dynamic_logic.dynamic_schema_documentation_construction import (
     get_restx_param_documentation,
 )
 from middleware.schema_and_dto_logic.non_dto_dataclasses import SchemaPopulateParameters
