+++ resources/ResetPassword.py
@@ -2,15 +2,13 @@ from flask import Response
 
 from middleware.primary_resource_logic.reset_token_queries import (
     reset_password,
-    RequestResetPasswordRequest,
-    RequestResetPasswordRequestSchema,
     ResetPasswordSchema,
     ResetPasswordDTO,
 )
 from utilities.namespace import create_namespace
 
 from resources.PsycopgResource import PsycopgResource, handle_exceptions
-from middleware.schema_and_dto_logic.dynamic_schema_documentation_construction import (
+from middleware.schema_and_dto_logic.dynamic_logic.dynamic_schema_documentation_construction import (
     get_restx_param_documentation,
 )
 from middleware.schema_and_dto_logic.non_dto_dataclasses import SchemaPopulateParameters
